added new target Markdown + generalized render
