+++ lib/Markup/Render/MarkupJekyll.swift
@@ -7,13 +7,8 @@ public struct JekyllGenerator: Render {
         self.permalink = permalink
     }
 
-    public func render(content: String) -> String? {
-        let syntaxTree = SyntaxAnalyzer.parse(content: content)
-        guard syntaxTree.count > 0 else { return nil }
-
-        let filteredSyntaxTree = syntaxTree.filter { !$0.isHidden }.reduce()
-        filteredSyntaxTree.forEach { print($0) }
-        return filteredSyntaxTree.reduce("") { (acc, node) in acc + node.jekyll(permalink: permalink) }
+    func render(node: Node) -> String {
+        return node.jekyll(permalink: permalink)
     }
 }
