+++ markdown/Markdown.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:JekyllMarkdown.xcodeproj">
+      location = "self:/Users/miguelangel/Documents/3.projects/47deg/nef/markdown/Markdown.xcodeproj">
    </FileRef>
 </Workspace>
